+++ openslides/core/projector.py
@@ -10,19 +10,19 @@ from ..utils.projector import register_projector_element
 
 
 def countdown(
-    config: Dict[str, Any], all_data: Dict[str, Dict[int, Dict[str, Any]]]
+    element: Dict[str, Any], all_data: Dict[str, Dict[int, Dict[str, Any]]]
 ) -> Dict[str, Any]:
     """
     Countdown slide.
 
     Returns the full_data of the countdown element.
 
-    config = {
+    element = {
         name: 'core/countdown',
         id: 5,  # Countdown ID
     }
     """
-    countdown_id = config.get("id") or 1
+    countdown_id = element.get("id") or 1
 
     try:
         return all_data["core/countdown"][countdown_id]
@@ -31,19 +31,19 @@ def countdown(
 
 
 def message(
-    config: Dict[str, Any], all_data: Dict[str, Dict[int, Dict[str, Any]]]
+    element: Dict[str, Any], all_data: Dict[str, Dict[int, Dict[str, Any]]]
 ) -> Dict[str, Any]:
     """
     Message slide.
 
     Returns the full_data of the message element.
 
-    config = {
+    element = {
         name: 'core/projector-message',
         id: 5,  # ProjectorMessage ID
     }
     """
-    message_id = config.get("id") or 1
+    message_id = element.get("id") or 1
 
     try:
         return all_data["core/projector-message"][message_id]
@@ -54,4 +54,4 @@ def message(
 def register_projector_elements() -> None:
     register_projector_element("core/countdown", countdown)
     register_projector_element("core/projector-message", message)
-    # TODO: Deside if we need a clock slide
+    # TODO: Add clock slide