+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Core:
  - Switch from Yarn back to npm [#3964].
  - Added password reset link (password reset via email) [#3914].
  - Added global history mode [#3977].
+ - Projector Refactor [4119, #4130].
 
 Agenda:
  - Added viewpoint to assign multiple items to a new parent item [#4037].